Adding files via upload
Updated code in game_logic.cpp and main.cpp. The cells of the tic-tac-toe board are now labelled with numbers 1-9, matching the layout of a numpad. Original code is commented out so it can be used depending on user preference. Added feature that lets users know whose turn it is.